backend/requirements.txt
Django==3.2.16
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
django-filter==22.1
requests==2.26.0

flake8==3.9.2
flake8-broken-line
flake8-return
pep8-naming